Add emission selection
The feature added here is that you can select an emission liquid in an
automatic emitter with two dropdowns (type and specific liquid).

However, the more important change is in the issue that came to light
during the process of implementing this feature: that there is necessity
for a generic Liquid from which Base and Mixture should inherit.

The namespace Liquid is now LiquidHandling to avoid confusion with the
new Liquid class, which, again, is the parent class for both Base and
Mixture.

Additionally, it is now possible to add any number of Liquid subtypes,
not that I truly think we will ever need that architecture. But I rest
better knowing that it's just done the "right" way.